Fixed more GCC warnings, alot of 64bit portability things mostly. Minor PNG header fix. Changed instances of errno to strerror(errno).
git-svn-id: https://svn.code.sf.net/p/fteqw/code/branches/wip@3805 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/com_mesh.c
@@ -2391,7 +2391,7 @@ qboolean Mod_LoadQ1Model (model_t *mod, void *buffer)
 	version = LittleLong(pq1inmodel->version);
 	if (version == QTESTALIAS_VERSION)
 	{
-		hdrsize = (unsigned int)&((dmdl_t*)NULL)->flags;
+		hdrsize = (size_t)&((dmdl_t*)NULL)->flags;
 		qtest = true;
 	}
 	else if (version == 50)
